Adds 4 more plugins and upgrades the install script

+++ install.sh
@@ -1,16 +1,47 @@
 #!/bin/sh
 
-rm -rf $(pwd)/bundle/Vundle.vim/
-#Works like an if
-#[ -d ~/.vim/bundle/Vundle.vim ] || git clone https://github.com/gmarik/Vundle.vim.git ~/.vim/bundle/Vundle.vim
-git clone https://github.com/gmarik/Vundle.vim.git ~/.vim/bundle/Vundle.vim
+if [ "$(ls -A $(pwd)/bundle/Vundle.vim/)" ]; then
+        echo "Vundle already installed!"
+else
+        rm -rf $(pwd)/bundle/Vundle.vim/
+        git clone https://github.com/gmarik/Vundle.vim.git ~/.vim/bundle/Vundle.vim
+fi
+
+ln -sf $(pwd)/vimrc ~/.vimrc
+echo "vimrc symlinked"
+
+vim +PluginInstall +PluginClean +qall
+echo "Plugins installed!"
+
+echo "Setting oh my zsh..."
 [ -d ~/.oh-my-zsh ] || git clone git://github.com/robbyrussell/oh-my-zsh.git ~/.oh-my-zsh
 [ -d ~/.oh-my-zsh/custom/plugins/zsh-syntax-highlighting ] || git clone git://github.com/zsh-users/zsh-syntax-highlighting.git ~/.oh-my-zsh/custom/plugins/zsh-syntax-highlighting
-ln -sf $(pwd)/vimrc ~/.vimrc
-ln -sf $(pwd)/Xdefaults ~/.Xdefaults
+echo "oh my zsh configured!"
+
 ln -sf $(pwd)/zshrc ~/.zshrc
-ln -sf $(pwd)/dwmstart /usr/share/xsessions/dwmstart
-vim +PluginInstall +PluginClean +qall
+echo "zshrc symlinked"
+
+echo "Do you want to use the custom dwmstart? (y/n)"
+read DWM
+
+if [ "$DWM" = "y" ]
+then
+        ln -sf $(pwd)/dwmstart /usr/share/xsessions/dwmstart
+        echo "dwmstart symlinked"
+fi
+
+echo "Do you want to use the custom Xdefaults? (y/n)"
+read XDEF
+
+if [ "$XDEF" = "y" ]
+then
+        ln -sf $(pwd)/Xdefaults ~/.Xdefaults
+        echo ".Xdefaults symlinked"
+fi
+
+echo "Compiling vimproc..."
 cd ~/.vim/bundle/vimproc.vim
 make
+echo "vimproc compiled!"
+
 echo "Don't forget to add dwmstart to execute"
